changes in my working copy

+++ src/buildutil/testsuite/test-Buffer.cc
@@ -30,8 +30,8 @@ TEST(PedigreeBuffer, InitialSettings)
 {
     Buffer<uint8_t> buffer(32768);
 
-    EXPECT_EQ(buffer.getDataSize(), 0);
-    EXPECT_EQ(buffer.getSize(), 32768);
+    EXPECT_EQ(buffer.getDataSize(), 0U);
+    EXPECT_EQ(buffer.getSize(), 32768U);
 }
 
 TEST(PedigreeBuffer, ReadEmpty)
@@ -43,7 +43,7 @@ TEST(PedigreeBuffer, ReadEmpty)
 
     size_t r = buffer.read(buf, 16, false);
 
-    EXPECT_EQ(r, 0);
+    EXPECT_EQ(r, 0U);
 }
 
 TEST(PedigreeBuffer, TooManyWithShort)
@@ -53,8 +53,8 @@ TEST(PedigreeBuffer, TooManyWithShort)
     char buf[16];
     memset(buf, 0, 16);
 
-    EXPECT_EQ(buffer.write(buf, 16, false), 8);
-    EXPECT_EQ(buffer.read(buf, 16, false), 8);
+    EXPECT_EQ(buffer.write(buf, 16, false), 8U);
+    EXPECT_EQ(buffer.read(buf, 16, false), 8U);
 }
 
 TEST(PedigreeBuffer, TooManyWithoutShort)
@@ -64,8 +64,8 @@ TEST(PedigreeBuffer, TooManyWithoutShort)
     char buf[16];
     memset(buf, 0, 16);
 
-    EXPECT_EQ(buffer.write(buf, 16, false), 8);
-    EXPECT_EQ(buffer.read(buf, 16, false), 8);
+    EXPECT_EQ(buffer.write(buf, 16, false), 8U);
+    EXPECT_EQ(buffer.read(buf, 16, false), 8U);
 }
 
 TEST(PedigreeBuffer, ReadTooMany)
@@ -75,8 +75,8 @@ TEST(PedigreeBuffer, ReadTooMany)
     char buf[16];
     memset(buf, 0, 16);
 
-    EXPECT_EQ(buffer.write(buf, 6, false), 6);
-    EXPECT_EQ(buffer.read(buf, 8, false), 6);
+    EXPECT_EQ(buffer.write(buf, 6, false), 6U);
+    EXPECT_EQ(buffer.read(buf, 8, false), 6U);
 }
 
 TEST(PedigreeBuffer, Overflow)
@@ -87,8 +87,8 @@ TEST(PedigreeBuffer, Overflow)
     memset(buf, 0xAB, 16);
 
     // Overflow truncates.
-    EXPECT_EQ(buffer.write(buf, 16, false), 8);
-    EXPECT_EQ(buffer.getDataSize(), 8);
+    EXPECT_EQ(buffer.write(buf, 16, false), 8U);
+    EXPECT_EQ(buffer.getDataSize(), 8U);
 }
 
 TEST(PedigreeBuffer, OverlapReader)
@@ -98,9 +98,9 @@ TEST(PedigreeBuffer, OverlapReader)
     char buf[16], buf2[16];
     memset(buf, 0xAB, 16);
 
-    EXPECT_EQ(buffer.write(buf, 6, false), 6);
-    EXPECT_EQ(buffer.read(buf2, 2, false), 2);
-    EXPECT_EQ(buffer.write(buf, 4, false), 4);
+    EXPECT_EQ(buffer.write(buf, 6, false), 6U);
+    EXPECT_EQ(buffer.read(buf2, 2, false), 2U);
+    EXPECT_EQ(buffer.write(buf, 4, false), 4U);
 }
 
 TEST(PedigreeBuffer, Overlap)
@@ -110,13 +110,13 @@ TEST(PedigreeBuffer, Overlap)
     char buf[16], buf2[16];
     memset(buf, 0xAB, 16);
 
-    EXPECT_EQ(buffer.write(buf, 6, false), 6);
-    EXPECT_EQ(buffer.read(buf2, 6, false), 6);
+    EXPECT_EQ(buffer.write(buf, 6, false), 6U);
+    EXPECT_EQ(buffer.read(buf2, 6, false), 6U);
     EXPECT_TRUE(memcmp(buf, buf2, 6) == 0);
 
     // Rolls over - two bytes at the end and then two bytes at the start.
-    EXPECT_EQ(buffer.write(buf, 4, false), 4);
-    EXPECT_EQ(buffer.read(buf2, 4, false), 4);
+    EXPECT_EQ(buffer.write(buf, 4, false), 4U);
+    EXPECT_EQ(buffer.read(buf2, 4, false), 4U);
 
     // Verify the reader can catch up.
     EXPECT_TRUE(memcmp(buf, buf2, 4) == 0);
@@ -138,8 +138,8 @@ TEST(PedigreeBuffer, FillBuffer)
     }
 
     // Can't write to a full buffer.
-    EXPECT_EQ(buffer.write(buf, 16, false), 0);
-    EXPECT_EQ(buffer.getDataSize(), 32768);
+    EXPECT_EQ(buffer.write(buf, 16, false), 0U);
+    EXPECT_EQ(buffer.getDataSize(), 32768U);
 }
 
 TEST(PedigreeBuffer, FillAndRead)
@@ -151,10 +151,10 @@ TEST(PedigreeBuffer, FillAndRead)
     memset(out, 0, 32768);
 
     size_t r = buffer.write(buf, 32768, false);
-    EXPECT_EQ(r, 32768);
+    EXPECT_EQ(r, 32768U);
 
     r = buffer.read(out, 32768, false);
-    EXPECT_EQ(r, 32768);
+    EXPECT_EQ(r, 32768U);
 
     EXPECT_TRUE(memcmp(buf, out, 32768) == 0);
 }
@@ -163,15 +163,15 @@ TEST(PedigreeBuffer, Chase)
 {
     Buffer<size_t> buffer(32768);
 
-    const int n = 0x10000;
-    const int readThreshold = 0x500;
+    const size_t n = 0x10000;
+    const size_t readThreshold = 0x500;
 
     std::unique_ptr<size_t[]> numbers(new size_t[n]);
     size_t offset = 0;
 
     for (size_t i = 0; i < n; ++i)
     {
-        ASSERT_EQ(buffer.write(&i, 1, false), 1);
+        ASSERT_EQ(buffer.write(&i, 1, false), 1U);
 
         if (i && (i % readThreshold == 0))
         {
